Move functions and constants into own files
- Add `functions.php` & `consts.php`, each with own namespace
- Move constants to `Conts` & functions to `Functions`
- `use` only what is necessary in `svglint.php`
